requirements.txt
asgiref==3.10.0
Django==5.1.14
djangorestframework==3.15.2
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
